Fixes Lab component to move experiment setup outside of the component
