+++ cdap-ui/app/cdap/components/Lab/types.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright © 2021 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+export interface IExperiment {
+  experimentId: string;
+  enabled: boolean;
+  screenshot: string | null;
+  name: string;
+  description: string;
+  showValue?: boolean;
+  valueLabel?: string;
+  valueType?: string;
+  force?: boolean;
+}
+export interface ILabState {
+  experiments: IExperiment[];
+}
